requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
transformers==4.44.2
torch==2.4.1
pypdf2==3.0.1
pdfplumber==0.11.4
redis==5.0.8
python-multipart==0.0.9
pytest==8.3.3
requests==2.31.0
python-dotenv==1.0.0